Fixing multi-start issue with $My being readonly
If the script is included more than once in an active session, the second and subsequent inclusion fails because the script tries to set a read-only variable $My which has already been set in the first include. The solution is to just force the update and still retain read=only protect to avoid it being overwritten by some other script.